new bitmex support. Adds some tests now that its all in its own func. Fixes poloniex bug

+++ exchanges/poloniex/poloniex_types.go
@@ -8,7 +8,7 @@ import (
 
 // Ticker holds ticker data
 type Ticker struct {
-	ID            int64   `json:"id"`
+	ID            float64   `json:"id"`
 	Last          float64 `json:"last,string"`
 	LowestAsk     float64 `json:"lowestAsk,string"`
 	HighestBid    float64 `json:"highestBid,string"`
